// packages/game-calendar/src/client/firebase/models.ts
export interface RefreshResponse {
  expires_in: string
  token_type: 'Bearer'
  refresh_token: string
  id_token: string
  user_id: string
  project_id: string
}
// ...
export interface FirestoreResponse {
  documents?: FirestoreDocument[]
}

export interface User {
  email: string
  idToken: string
  localId: string
  refreshToken: string
}

export interface FirestoreDocument extends MapValueInternal {
  name: string
  createTime: string
  updateTime: string
}

export interface ArrayValue {
  arrayValue: ArrayValueInternal
}

export interface ArrayValueInternal {
  values?: Value[]
}

export interface MapValue {
  mapValue: MapValueInternal
}

export interface MapValueInternal {
  fields?: Record<string, Value>
}